requirements.txt
asgiref==3.7.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.1
django-dotenv==1.4.2
django-filter==23.5
djangorestframework==3.14.0
fonttools==4.47.2
future==0.18.3
kiwisolver==1.4.5
matplotlib==3.8.2
mysqlclient==2.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
response==0.5.0
scipy==1.12.0
serializers==0.2.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
